MainModule For Whitelist -- REVOKING NOT
This is a tip Only This script Is only used for the must people as a main module script as a require script
Like 
require(1234) -- Asset id
A Asset Script must be publish. After that you need click Publish. Then you get that ID That shown you. Then go on a script and do Require(12345) -- your asset id remove 12345